"Redirect users to their respective dashboards after deleting a user or reservation"
This commit includes changes to the delete\_Gebruiker.php and delete\_reservation.php controllers, as well as the overzichtreserveren.php template.

In the delete\_Gebruiker.php and delete\_reservation.php controllers, the redirect location now depends on the user's role. This is determined by checking the value of $_SESSION['userRole'] and using a switch statement to set the appropriate Location header.

In the overzichtreserveren.php template, the number of columns in the table has been increased from 4 to 6. This is unrelated to the changes in the controller files, but is included in the same commit for convenience.